Ensure code evaluated as middleware
Make sure that code running to do things such as defining now, today, two weeks ago, etc is run as middleware and set as a local variable (in res.locals). This will ensure it is reevaluated every time a page loads. OTHERWISE it will only be run when the app is (re)started!!! This was causing a lot of silent bugs.